Fixed #23 (soft hyphens buggy in Chrome)
Chrome did not report the position of soft hyphens, considering them as
whitespace, causing bugs. There was a mitigation in place to avoid this
bug, for for some reason a refactoring must have broken it. This is now
fixed. As a bonus, soft hyphens now display when a word is broken across
regions.