fixes #56: exception thrown during relayout after scrolling down
Note: Though I strive to keep as much code intact as possible, for this fix I changed `fixNullRect` to require the target rect as parameter and to return the (potentially) fixed rect, instead of changing the `rect` variable directly.
This allows this fix to remain concise and I think it also improves the readability of the code.