add multi-column support
this basically translates to: break content that overflows horizontally as well as vertically, instead of only checking vertical overflow.

This is a major change, since previously the alogorithm ignored horizontal overflow. With this patch in place, content will be also split where horizontal overflow is detected.
The primary reason is that with multi-columns activated in CSS, the content no longer overflows verticall, but instead overflows horizontally.